+++ packages/button/src/css/index.js
@@ -2,10 +2,10 @@ import core from '@pluralsight/ps-design-system-core'
 import {
   defaultName as themeDefaultName,
   names as themeNames
-} from '@pluralsight/ps-design-system-theme/vars'
-import { transparentize } from '@pluralsight/ps-design-system-util/color'
+} from '@pluralsight/ps-design-system-theme/vars.js'
+import { transparentize } from '@pluralsight/ps-design-system-util/color.js'
 
-import * as vars from '../vars'
+import * as vars from '../vars/index.js'
 
 export default {
   '@keyframes psds-button__keyframes__spin': {
@@ -29,32 +29,28 @@ export default {
     whiteSpace: 'nowrap',
     textDecoration: 'none',
     transition: `all ${core.motion.speedNormal}`,
-    verticalAlign: 'middle'
-  },
-  '.psds-button:hover': {
-    background: core.colors.orangeLight,
-    cursor: 'pointer'
-  },
-  '.psds-button:focus': {
-    outline: 'none'
+    verticalAlign: 'middle',
+
+    '&:not([disabled]):hover': {
+      background: core.colors.orangeLight,
+      cursor: 'pointer'
+    },
+    '&:focus': {
+      outline: 'none',
+
+      '&:before': {
+        content: ' ',
+        position: 'absolute',
+        top: '-4px',
+        left: '-4px',
+        right: '-4px',
+        bottom: '-4px',
+        border: `3px solid ${core.colors.blue}`,
+        borderRadius: '4px'
+      }
+    }
   },
 
-  // :focus
-  [`.psds-button.psds-theme--${themeNames.light}:focus`]: {
-    border: '1px solid transparent'
-  },
-  '.psds-button:focus:before': {
-    content: ' ',
-    position: 'absolute',
-    top: '-4px',
-    left: '-4px',
-    right: '-4px',
-    bottom: '-4px',
-    border: `3px solid ${core.colors.blue}`,
-    borderRadius: '4px'
-  },
-
-  // --size
   [`.psds-button--size-${vars.sizes.xSmall}`]: {
     fontSize: core.type.fontSizeXSmall,
     padding: `0 ${core.layout.spacingXSmall}`,
@@ -75,85 +71,79 @@ export default {
     padding: `${core.layout.spacingXXSmall} ${core.layout.spacingMedium}`,
     height: '48px'
   },
-  // --appearance
+
   [`.psds-button--appearance-${vars.appearances.secondary}`]: {
     color: core.colors.gray01,
-    background: transparentize(0.75, core.colors.gray02)
+    background: transparentize(0.75, core.colors.gray02),
+
+    '&:not([disabled]):hover': {
+      color: core.colors.white,
+      background: transparentize(0.65, core.colors.gray02)
+    }
   },
-  [`.psds-button--appearance-${vars.appearances.secondary}:hover`]: {
-    color: core.colors.white,
-    background: transparentize(0.65, core.colors.gray02)
-  },
-  [`.psds-button--appearance-${vars.appearances.secondary}.psds-theme--${
-    themeNames.light
-  }`]: {
+  [`.psds-button--appearance-${vars.appearances.secondary}.psds-theme--${themeNames.light}`]: {
     color: core.colors.gray03,
-    background: transparentize(0.65, core.colors.gray01)
-  },
-  [`.psds-button--appearance-${vars.appearances.secondary}.psds-theme--${
-    themeNames.light
-  }:hover`]: {
-    color: core.colors.gray06,
-    background: transparentize(0.35, core.colors.gray01)
+    background: transparentize(0.65, core.colors.gray01),
+
+    '&:not([disabled]):hover': {
+      color: core.colors.gray06,
+      background: transparentize(0.35, core.colors.gray01)
+    }
   },
   [`.psds-button--appearance-${vars.appearances.stroke}`]: {
     border: `1px solid ${core.colors.orange}`,
     color: core.colors.orange,
-    background: 'none'
-  },
-  [`.psds-button--appearance-${vars.appearances.stroke}:hover`]: {
-    border: `1px solid ${core.colors.orangeLight}`,
-    color: core.colors.orangeLight,
-    background: 'none'
+    background: 'none',
+
+    '&:not([disabled]):hover': {
+      border: `1px solid ${core.colors.orangeLight}`,
+      color: core.colors.orangeLight,
+      background: 'none'
+    }
   },
   [`.psds-button--appearance-${vars.appearances.flat}`]: {
     border: 'none',
-    background: 'none'
+    background: 'none',
+
+    '&:not([disabled]):hover': {
+      color: core.colors.white,
+      background: transparentize(0.85, core.colors.white)
+    }
   },
-  [`.psds-button--appearance-${vars.appearances.flat}.psds-theme--${
-    themeNames.light
-  }`]: {
-    color: core.colors.gray03
+  [`.psds-button--appearance-${vars.appearances.flat}.psds-theme--${themeNames.light}`]: {
+    color: core.colors.gray03,
+
+    '&:not([disabled]):hover': {
+      color: 'inherit',
+      background: transparentize(0.85, core.colors.gray03)
+    }
   },
-  [`.psds-button--appearance-${
-    vars.appearances.flat
-  }.psds-theme--${themeDefaultName}`]: {
+  [`.psds-button--appearance-${vars.appearances.flat}.psds-theme--${themeDefaultName}`]: {
     color: core.colors.gray02
   },
-  [`.psds-button--appearance-${vars.appearances.flat}.psds-theme--${
-    themeNames.light
-  }:hover`]: {
-    background: transparentize(0.85, core.colors.gray03)
-  },
-  [`.psds-button--appearance-${
-    vars.appearances.flat
-  }.psds-theme--${themeDefaultName}:hover`]: {
-    color: core.colors.white,
-    background: transparentize(0.85, core.colors.white)
-  },
-  // --disabled
+
   [`.psds-button--disabled`]: {
     color: core.colors.gray02,
     background: core.colors.gray03,
-    cursor: 'default'
+    cursor: 'default',
+
+    '&:not([disabled]):hover': {
+      color: 'inherit',
+      background: 'inherit',
+      border: 'inherit'
+    }
   },
   [`.psds-button--disabled.psds-theme--${themeNames.light}`]: {
     color: core.colors.gray03,
     background: core.colors.gray01
   },
-  [`.psds-button--disabled.psds-button--appearance-${
-    vars.appearances.primary
-  }`]: {
+  [`.psds-button--disabled.psds-button--appearance-${vars.appearances.primary}`]: {
     opacity: 0.5
   },
-  [`.psds-button--disabled.psds-button--appearance-${
-    vars.appearances.secondary
-  }`]: {
+  [`.psds-button--disabled.psds-button--appearance-${vars.appearances.secondary}`]: {
     opacity: 0.5
   },
-  [`.psds-button--disabled.psds-button--appearance-${
-    vars.appearances.stroke
-  }`]: {
+  [`.psds-button--disabled.psds-button--appearance-${vars.appearances.stroke}`]: {
     border: 'none',
     opacity: 0.5
   },
@@ -161,28 +151,20 @@ export default {
     opacity: 0.4,
     background: 'none'
   },
-  // --iconAlign / iconOnly
+
   [`.psds-button--iconAlign-${vars.iconAligns.right}`]: {
     flexDirection: 'row-reverse'
   },
-  [`.psds-button--iconAlign-${
-    vars.iconAligns.right
-  }.psds-button--not-iconOnly`]: {
+  [`.psds-button--iconAlign-${vars.iconAligns.right}.psds-button--not-iconOnly`]: {
     paddingRight: core.layout.spacingXSmall
   },
-  [`.psds-button--iconAlign-${
-    vars.iconAligns.left
-  }.psds-button--not-iconOnly`]: {
+  [`.psds-button--iconAlign-${vars.iconAligns.left}.psds-button--not-iconOnly`]: {
     paddingLeft: core.layout.spacingXSmall
   },
-  [`.psds-button--iconAlign-${
-    vars.iconAligns.right
-  }.psds-button--not-iconOnly.psds-button--size-${vars.sizes.large}`]: {
+  [`.psds-button--iconAlign-${vars.iconAligns.right}.psds-button--not-iconOnly.psds-button--size-${vars.sizes.large}`]: {
     paddingRight: core.layout.spacingSmall
   },
-  [`.psds-button--iconAlign-${
-    vars.iconAligns.left
-  }.psds-button--not-iconOnly.psds-button--size-${vars.sizes.large}`]: {
+  [`.psds-button--iconAlign-${vars.iconAligns.left}.psds-button--not-iconOnly.psds-button--size-${vars.sizes.large}`]: {
     paddingLeft: core.layout.spacingSmall
   },
   [`.psds-button--iconOnly`]: {
@@ -200,7 +182,7 @@ export default {
   [`.psds-button--iconOnly.psds-button--size-${vars.sizes.large}`]: {
     width: '48px'
   },
-  // __icon
+
   [`.psds-button__icon`]: {
     display: 'flex',
     alignItems: 'center',
@@ -219,7 +201,7 @@ export default {
     width: '100%',
     margin: 0
   },
-  // __loading
+
   [`.psds-button__loading`]: ({ spin }) => ({
     display: 'inline-block',
     height: 'calc(100% - 4px)',
@@ -238,9 +220,7 @@ export default {
     borderColor: transparentize(0.8, core.colors.gray01),
     borderTopColor: core.colors.gray01
   },
-  [`.psds-button__loading--appearance-${
-    vars.appearances.secondary
-  }.psds-button__loading--theme-${themeNames.light}`]: {
+  [`.psds-button__loading--appearance-${vars.appearances.secondary}.psds-button__loading--theme-${themeNames.light}`]: {
     borderColor: transparentize(0.8, core.colors.gray04),
     borderTopColor: core.colors.gray04
   },
@@ -248,19 +228,15 @@ export default {
     borderColor: transparentize(0.8, core.colors.white),
     borderTopColor: core.colors.orange
   },
-  [`.psds-button__loading--appearance-${
-    vars.appearances.flat
-  }.psds-button__loading--theme-${themeNames.light}`]: {
+  [`.psds-button__loading--appearance-${vars.appearances.flat}.psds-button__loading--theme-${themeNames.light}`]: {
     borderColor: transparentize(0.8, core.colors.gray04),
     borderTopColor: core.colors.gray04
   },
-  [`.psds-button__loading--appearance-${
-    vars.appearances.flat
-  }.psds-button__loading--theme-${themeDefaultName}`]: {
+  [`.psds-button__loading--appearance-${vars.appearances.flat}.psds-button__loading--theme-${themeDefaultName}`]: {
     borderColor: transparentize(0.8, core.colors.white),
     borderTopColor: core.colors.white
   },
-  // __text
+
   [`.psds-button__text`]: {
     alignItems: 'center',
     display: 'inline-flex',