+++ packages/dropdown/src/css/index.js
@@ -1,6 +1,6 @@
 import core from '@pluralsight/ps-design-system-core'
-import * as iconVars from '@pluralsight/ps-design-system-icon/vars'
-import { names as themeNames } from '@pluralsight/ps-design-system-theme/vars'
+import * as iconVars from '@pluralsight/ps-design-system-icon/vars.js'
+import { names as themeNames } from '@pluralsight/ps-design-system-theme/vars.js'
 
 import * as vars from '../vars/index.js'
 
@@ -24,23 +24,23 @@ export default {
     lineHeight: core.type.lineHeightStandard,
     fontWeight: core.type.fontWeightMedium,
     color: core.colors.gray03,
-    padding: `0 calc(${core.layout.spacingXSmall} + ${
-      iconVars.widths.medium
-    } + ${core.layout.spacingXSmall}) 0 ${core.layout.spacingMedium}`,
+    padding: `0 calc(${core.layout.spacingXSmall} + ${iconVars.widths.medium} + ${core.layout.spacingXSmall}) 0 ${core.layout.spacingMedium}`,
     border: 'none',
     textAlign: 'left',
-    cursor: 'pointer'
-  },
-  '.psds-dropdown__field:focus': {
-    outline: 'none'
-  },
-  [`.psds-dropdown__field.psds-theme--${themeNames.light}:focus`]: {
-    border: '1px solid transparent'
+    cursor: 'pointer',
+
+    '&:focus': {
+      outline: 'none'
+    }
   },
   [`.psds-dropdown__field.psds-theme--${themeNames.light}`]: {
     background: core.colors.white,
     color: core.colors.gray05,
-    border: `1px solid ${core.colors.gray02}`
+    border: `1px solid ${core.colors.gray02}`,
+
+    '&:focus': {
+      border: '1px solid transparent'
+    }
   },
   [`.psds-dropdown__field--error.psds-theme--${themeNames.light}`]: {
     border: '1px solid transparent'
@@ -109,9 +109,7 @@ export default {
   // __error
   '.psds-dropdown__error': {
     position: 'absolute',
-    right: `calc(-1 * (${iconVars.widths.medium} + ${
-      core.layout.spacingXSmall
-    }))`,
+    right: `calc(-1 * (${iconVars.widths.medium} + ${core.layout.spacingXSmall}))`,
     display: 'flex',
     alignItems: 'center',
     color: core.colors.red,