+++ packages/tab/src/react/__specs__/storyshots.spec.js
@@ -1,0 +1,12 @@
+import initStoryshots, {
+  snapshotWithOptions
+} from '@storybook/addon-storyshots'
+
+jest.mock('@pluralsight/ps-design-system-storybook-addon-theme')
+
+const createNodeMock = el => document.createElement('div')
+
+initStoryshots({
+  framework: 'react',
+  test: snapshotWithOptions({ createNodeMock })
+})